+++ src/entries/options.js
@@ -92,6 +92,7 @@ function renderAccounts (accounts) {
       Account.create({type: 'nextcloud', url: 'http://example.org', username: 'bob', password: 'password'})
         .then(() => triggerRender())
     }}>Add account</a>
+    <a className="test-link" href="./test.html">run tests</a>
   </div>
 }
 function renderPicker (cb, tree) {